reafactor createStore with zustand
